refactor : ai 도메인 분리
