+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/ai/dto/AIRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.dto;
+package com.sparta.blackwhitedeliverydriver.ai.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
